requirements.txt
python-dotenv==1.0.1
langchain==0.1.16
langchain_community==0.0.34
langchain_openai==0.1.3
unstructured==0.13.3